requirements.txt
flask
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pydantic
websockets==13.1